chore: f64 -> f32

+++ src/consts.rs
@@ -5,7 +5,7 @@ pub const TICK_RATE_IN_MILI: u64 = 33;
 
 // Not sure why rustc is warning about these constants being unused
 #[allow(dead_code)]
-pub const FRAMES_PER_SECOND: f64 = 1000.0 / TICK_RATE_IN_MILI as f64;
+pub const FRAMES_PER_SECOND: f32 = 1000.0 / TICK_RATE_IN_MILI as f32;
 
 #[allow(dead_code)]
 pub const NUM_PARTICLES: usize = 75;
@@ -14,7 +14,7 @@ pub const NUM_PARTICLES: usize = 75;
 pub const CHARACTERS: [char; 6] = ['█', '▓', '▒', '░', '▄', '▀'];
 
 #[allow(dead_code)]
-pub const TERMINAL_GRAVITY: Vector2<f64> = Vector2::new(0.0, 9.81);
+pub const TERMINAL_GRAVITY: Vector2<f32> = Vector2::new(0.0, 9.81);
 
 #[allow(dead_code)]
 pub const COLORS: [Color; 5] = [
@@ -26,4 +26,4 @@ pub const COLORS: [Color; 5] = [
 ];
 
 //
-// pub const GRAVITY: Vector3<f64> = Vector3::new(0.0, -9.81, 0.0);
+// pub const GRAVITY: Vector3<f32> = Vector3::new(0.0, -9.81, 0.0);
